requirements.txt
streamlit==1.56.0

torch>=2.0.0,<2.3.0
torchvision>=0.15.0,<0.18.0

transformers>=4.40.0,<4.45.0
pillow>=10.0.0

numpy>=1.24.0
opencv-python>=4.8.0
easyocr>=1.7.1

pdf2image>=1.16.3
pypdfium2>=4.30.0
PyPDF2>=3.0.1

pydantic>=2.0.0,<3.0.0

matplotlib>=3.7.0
